Added show method for genambig, Genotypes
The original version of Genotypes produced an awkwardly formatted matrix,
with heterozygous loci reported as 'integer 2'. I modified the method so
that it actually shows a character representation of the actual
genotype (e.g., "240/246/252").

I also added a `show` method for class genambig. This adds a bit of
formatting to the display to make it a little easier to parse when viewed
at the console.